Fix heavy statistics.
The major collector refactoring left heavy statistics broken.  A few
heavy statistics counters are defined in sgen-gc.c but used in the
major collectors.  We're now exporting them.  Another possibility
would have been to define them in both major collectors, but since
this is only a debug feature that's not normally compiled in, we don't
care.