[sgen] Fixed heap for Mark&Sweep.
This is a new variant of Mark&Sweep which uses a pre-allocated
fixed-size heap instead of a dynamically grown, segmented one.  The
main advantage of this heap organization is that it takes only one
pointer comparison to determine whether an object is on the small
object heap or in the LOS, instead of inspecting the object's size,
which requires three loads.  Since that determination is required for
every marked object, speeding it up can improve performance
significantly.