DEBUG: Check that thread info list doesn't change during GC.
One of my hypotheses was that the thread info list is added to during
GCs by the threads that are still starting up and have hence not been
paused by STW yet.  That doesn't happen - I guess that's what the
thread suspend semaphore is for.

It is the case, however, that at the point at which the program fails
the thread info list has changed from what it was during the last GC
pause.