DEBUG: Remember last stopped IPs of threads.
I thought I'd take a look at where STW interrupts threads in the crash
cases, so I save the stopped IPs from the last 4 stops.

This turns out to be quite interesting.  In the thread that fails with
the assertion, we always get a stop somewhere in the code associated
with the LLS.  Here's a few examples:

#8  0x00000000006ca3c3 in sgen_thread_detach (p=0x7fffe405f280) at sgen-gc.c:4159
4159			mono_thread_detach (mono_thread_current ());
(gdb) p p->last_stopped_ip
$60 = {0x74c10d <mono_memory_write_barrier+9>, 0x0, 0x0, 0x0}
$61 = {0x74c0f7 <mono_memory_barrier+7>, 0x0, 0x0, 0x0}
$62 = {0x74c6ed <mono_thread_info_lookup+76>, 0x0, 0x0, 0x0}
$63 = {0x7ffff71cf75d <write+45>, 0x74c0f8 <mono_memory_barrier+8>,
  0x74c0f8 <mono_memory_barrier+8>, 0x74c0f8 <mono_memory_barrier+8>}
$64 = {0x74c20d <get_hazardous_pointer_with_mask+162>, 0x0, 0x0, 0x0}
$65 = {0x7ffff71cf75d <write+45>, 0x74c312 <mono_lls_find+115>,
  0x74c312 <mono_lls_find+115>, 0x74c312 <mono_lls_find+115>}
$66 = {0x74c104 <mono_memory_write_barrier>, 0x7ffff71d8d47 <mprotect+7>,
  0x7ffff71d8d47 <mprotect+7>, 0x7ffff71d8d47 <mprotect+7>}
$67 = {0x74c1fc <get_hazardous_pointer_with_mask+145>, 0x0, 0x0, 0x0}

Unfortunately they're all over the place.  Note the cases where the
last stop is in write() - I'm guessing that's because the bug occurred
in the stop before that and then the thread is interrupted again when
it's printing out the assertion failure message.

I also looked at the stopped IPs for the other threads, and there
didn't seem to be an apparent pattern, but it's worth looking into
that again.